requirements.txt
Flask
redis
grpcio
googleapis-common-protos